# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Ecommerce",
    "version": "1.0.0",
    "description": "Documentación de la API para la tienda Ecommerce"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único del usuario",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Juan Pérez"
          },
          "correo": {
            "type": "string",
            "description": "Correo electrónico del usuario",
            "example": "juan.perez@example.com"
          },
          "rol": {
            "type": "string",
            "description": "Rol del usuario (ej. cliente, administrador)",
            "example": "cliente"
          },
          "fechaRegistro": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de registro del usuario",
            "example": "2023-10-01T12:34:56.789Z"
          }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único del producto",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Smartphone XYZ"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción del producto",
            "example": "Smartphone de última generación con múltiples funcionalidades."
          },
          "precio": {
            "type": "number",
            "format": "float",
            "description": "Precio del producto",
            "example": 299.99
          },
          "cantidadEnStock": {
            "type": "integer",
            "description": "Cantidad disponible en stock",
            "example": 50
          },
          "categoriaId": {
            "type": "integer",
            "description": "ID de la categoría a la que pertenece el producto",
            "example": 1
          },
          "imagenes": {
            "type": "array",
            "items": {
              "type": "string",
              "description": "Nombre de archivo de las imágenes del producto",
              "example": "smartphone1.jpg"
            }
          },
          "featured": {
            "type": "boolean",
            "description": "Indica si el producto es destacado",
            "example": true
          }
        }
      },
...
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Iniciar sesión y obtener un token JWT",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "correo",
                  "contraseña"
                ],
                "properties": {
                  "correo": {
                    "type": "string",
                    "description": "Correo electrónico del usuario"
                  },
                  "contraseña": {
                    "type": "string",
                    "description": "Contraseña del usuario"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token JWT generado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Usuario no encontrado o contraseña incorrecta"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Obtener los datos del usuario autenticado",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/carts": {
      "get": {
        "summary": "Obtener el carrito del usuario",
        "tags": [
          "Carts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del carrito",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cartItems": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "cantidad": {
                            "type": "integer"
                          },
                          "product": {
                            "$ref": "#/components/schemas/Product"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener el carrito"
          }
        }
      }
    },
    "/carts/add": {
      "post": {
        "summary": "Agregar un producto al carrito",
        "tags": [
          "Carts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productoId",
                  "cantidad"
                ],
                "properties": {
                  "productoId": {
                    "type": "integer",
                    "description": "ID del producto a agregar"
                  },
                  "cantidad": {
                    "type": "integer",
                    "description": "Cantidad del producto"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Producto agregado al carrito",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "cartItem": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "carritoId": {
                          "type": "integer"
                        },
                        "productoId": {
                          "type": "integer"
                        },
                        "cantidad": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Producto o cantidad inválida"
          },
          "500": {
            "description": "Error al agregar al carrito"
          }
        }
      }
    },
    "/carts/update": {
      "put": {
        "summary": "Actualizar la cantidad de un producto en el carrito",
        "tags": [
          "Carts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productoId",
                  "cantidad"
                ],
                "properties": {
                  "productoId": {
                    "type": "integer",
                    "description": "ID del producto a actualizar"
                  },
                  "cantidad": {
                    "type": "integer",
                    "description": "Nueva cantidad del producto"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cantidad actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "cartItem": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "cantidad": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Producto o cantidad inválida"
          },
          "404": {
            "description": "Producto no encontrado en el carrito"
          },
          "500": {
            "description": "Error al actualizar la cantidad"
          }
        }
      }
    },
    "/carts/remove/{productoId}": {
      "delete": {
        "summary": "Eliminar un producto del carrito",
        "tags": [
          "Carts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productoId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del producto a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Producto eliminado del carrito"
          },
          "404": {
            "description": "Producto no encontrado en el carrito"
          },
          "500": {
            "description": "Error al eliminar el producto"
          }
        }
      }
    },
    "/carts/clear": {
      "post": {
        "summary": "Limpiar el carrito completo",
        "tags": [
          "Carts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Carrito limpiado exitosamente"
          },
          "500": {
            "description": "Error al limpiar el carrito"
          }
        }
      }
    },
...
  },
...
}